% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{getColorPaletteFromNameContinuous}
\alias{getColorPaletteFromNameContinuous}
\title{Determine point size automatically}
\usage{
getColorPaletteFromNameContinuous(palette.name = "YlOrRd")
}
\arguments{
\item{palette.name}{name of palette often used, such as "YlOrRd". (default: "YlOrRd)}
}
\value{
character vector contain the colors
}
\description{
Determine point size automatically
}
