% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDensity2D}
\alias{plotDensity2D}
\title{Wrap for plotting 2D density}
\usage{
plotDensity2D(x, peaks)
}
\arguments{
\item{x}{matrix or data.frame; map data, row for sample, column for dimension}

\item{peaks}{integer or character; index or names of the peaks. (default: NULL)}
}
\description{
Wrap for plotting 2D density
}
\details{
use ks::kde for density estimation
}
