% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotMatrix.simple}
\alias{plotMatrix.simple}
\title{plot matrix (typically genes expression)}
\usage{
plotMatrix.simple(
  dat,
  out.prefix = NULL,
  mytitle = "Heatmap",
  show.number = NULL,
  gp.show.number = gpar(),
  my.cell_fun = NULL,
  do.clust = NULL,
  z.lo = NULL,
  z.hi = NULL,
  z.len = 100,
  palatte = NULL,
  clust.row = FALSE,
  clust.column = FALSE,
  show.dendrogram = FALSE,
  waterfall.row = FALSE,
  waterfall.column = FALSE,
  row.ann.dat = NULL,
  row.split = NULL,
  column.split = NULL,
  returnHT = FALSE,
  par.legend = list(),
  par.heatmap = list(),
  col.ht = NULL,
  col.ann = NULL,
  top_annotation = NULL,
  par.warterfall = list(score.alpha = 1.5, do.norm = T),
  pdf.width = 8,
  pdf.height = 8,
  fig.type = "pdf",
  exp.name = "Count",
  ...
)
}
\arguments{
\item{dat}{matrix; matrix}

\item{out.prefix}{character; output prefix.}

\item{mytitle}{character; (default: "Heatmap")}

\item{show.number}{logical; (default: NULL)}

\item{gp.show.number}{object of gpar; (default: gpar() )}

\item{my.cell_fun}{function; cell_fun of Heatmap (default: NULL )}

\item{do.clust}{logical, character or dendrogram; passed to both cluster_columns and cluster_rows of Heatmap. Higher priority than clust.row and clust.column (default: NULL)}

\item{z.lo}{double; (default: NULL)}

\item{z.hi}{double; (default: NULL)}

\item{z.len}{integer; (default: 100)}

\item{palatte}{character; (default: NULL)}

\item{clust.row}{logical, character or dendrogram; passed to cluster_rows of Heatmap (default: FALSE)}

\item{clust.column}{logical, character or dendrogram; passed to cluster_columns of Heatmap (default: FALSE)}

\item{show.dendrogram}{logical, whetehr show the dendrogram (default: FALSE)}

\item{waterfall.row}{logical, order rows to make plot like waterfall (default: FALSE)}

\item{waterfall.column}{logical, order rows to make plot like waterfall (default: FALSE)}

\item{row.ann.dat}{data.frame; data for row annotation; rownames(row.ann.dat) cannot be NULL; (default: NULL)}

\item{row.split}{vector; used for row; (default: NULL)}

\item{column.split}{vector; used for column; (default: NULL)}

\item{returnHT}{logical; whether return HT; (default: FALSE)}

\item{par.legend}{list; lengend parameters, used to overwrite the default setting; (default: list())}

\item{par.heatmap}{list; other heatmap parameters, (default: list())}

\item{col.ht}{vector; (default: NULL)}

\item{col.ann}{list; (default: NULL)}

\item{top_annotation}{passed to Heatmap; (default: NULL)}

\item{par.warterfall}{list; parameters for warterfall, sucah as score.alpha}

\item{pdf.width}{double; width of the output plot (default: 22)}

\item{pdf.height}{double; height of the output plot (default: 22)}

\item{fig.type}{character; type of output file (default: "pdf")}

\item{exp.name}{character; showd in the legend (default: "Count")}

\item{...}{parameters passed to ComplexHeatmap::draw}
}
\description{
plot matrix (typically genes expression)
}
\details{
plot matrix
}
