% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run.cutreeDynamic}
\alias{run.cutreeDynamic}
\title{run dynamicTreeCut::cutreeDynamic on rows of the given matrix}
\usage{
run.cutreeDynamic(
  dat,
  method.hclust = "ward.D2",
  method.distance = "spearman",
  ...
)
}
\arguments{
\item{dat}{data frame or matrix;}

\item{method.hclust}{character; clustering method for hclust [default: "ward.D2"]}

\item{method.distance}{character; distance method for hclust [default: "spearman"]}

\item{...}{parameters passed to dynamicTreeCut::cutreeDynamic}
}
\value{
a matrix with dimention as input ( samples in rows and variables in columns)
}
\description{
run dynamicTreeCut::cutreeDynamic on rows of the given matrix
}
\details{
dynamicTreeCut::cutreeDynamic on rows of the given matrix
}
