% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssc.average.gene}
\alias{ssc.average.gene}
\title{calcualte the average expression of genes of each cell}
\usage{
ssc.average.gene(
  obj,
  assay.name = "exprs",
  gene = NULL,
  ncell.downsample = NULL,
  avg = "mean",
  ret.type = "sce"
)
}
\arguments{
\item{obj}{object of \code{singleCellExperiment} class}

\item{assay.name}{character; which assay (default: "exprs")}

\item{gene}{list; named list; NULL for all genes (default: NULL)}

\item{ncell.downsample}{integer; for each group, number of cells downsample to. (default: NULL)}

\item{avg}{character; average method. can be one of "mean", "diff", "zscore" . (default: "mean")}

\item{ret.type}{character; return type. can be one of "data.melt", "data.cast", "data.mtx". (default: "data.melt")}
}
\description{
calcualte the average expression of genes of each cell
}
\details{
multiple average methods are implemented
}
