% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDotPlotFromGeneTable}
\alias{plotDotPlotFromGeneTable}
\title{plot a dotplot}
\usage{
plotDotPlotFromGeneTable(
  gene.tb,
  group.tb = NULL,
  out.prefix = NULL,
  mcls2Name = NULL,
  column.x = "meta.cluster",
  column.v = "comb.ES",
  column.size = NULL,
  clamp.v = c(-0.25, 0.5),
  clamp.size = c(-0.25, 0.5),
  func.scale.color = scale_colour_distiller,
  point.shape = 16,
  recipe = NULL,
  par.color = list(palette = "RdYlBu", breaks = c(-0.25, 0, 0.25, 0.5), direction = -1),
  par.size = list(breaks = c(-0.25, 0, 0.25, 0.5), range = c(0.2, 6), labels = c(-0.25,
    0, 0.25, 0.5), limits = c(-0.25, 0.5) * 1)
)
}
\arguments{
\item{gene.tb}{data.table;}

\item{group.tb}{data.table; columns including "geneID" and "Group" are required (default NULL)}

\item{out.prefix}{character; output prefix. (default NULL)}

\item{mcls2Name}{vector; cluster name conversion. (default NULL)}

\item{column.x}{character; which column used for x. (default "meta.cluster")}

\item{column.v}{character; which column used for color. (default "comb.ES")}

\item{column.size}{character; which column used for dot size. (default NULL, same for column.v)}

\item{clamp.v}{double vector; values for "v" will be clamped to the range defined by this parameter, such as c(-0.25,0.5). (default: c(-0.25,0.5) )}

\item{clamp.size}{double vector; values for "size" will be clamped to the range defined by this parameter, such as c(-0.25,0.5). (default: NULL )}

\item{func.scale.color}{function; scale_colour_xx. (default: scale_colour_distiller )}

\item{point.shape}{integer; point shape. (default: 16 )}

\item{recipe}{character; predefined recipe, such as "z.freq". (default: NULL )}

\item{par.color}{list; parameters for scale_colour_distiller.}

\item{par.size}{list; parameters for scale_size.}
}
\description{
plot a dotplot
}
\details{
plot a dotplot
}
